Move IsPrism into the internal module.
